Extract the results grid into ResultsGridPanel UserControl
Finishes the MainWindow decomposition (UI design rule 7). The results
DataGrid and everything only about it moves out of MainWindow into a
focused UserControl, the last big cluster after SchemaTreePanel (#155)
and QueryEditorPanel (#156).

What moved into ResultsGridPanel:
- The results card markup (grid + empty state + plan view + script-section
  chips) and the results DataGrid's context menu.
- The cell inspector overlay and its JSON editor, deliberately left behind
  by #156 — plus all of its code-behind: the manual JSON-editor two-way
  sync, JSON highlighting load/apply, and the theme rewrite (which the
  panel now drives off its own ActualThemeVariantChanged, like
  QueryEditorPanel does for the SQL editor).
- All grid code-behind: type-aware column building, inline cell editing and
  the staged-edit commit path, safe-mode dirty-row washes, header-click
  sorting, follow-FK navigation, copy, and export/import orchestration.

Like QueryEditorPanel, ResultsGridPanel inherits the window's MainViewModel
DataContext (it needs the active tab's Rows/EditContext, the FK cache, the
CellInspector VM, the Importer) and tracks the active tab off its own
DataContextChanged. MainWindow's _queryViewModel, AttachQuery, and
OnViewModelPropertyChanged are gone — the grid was their last consumer.

The Export/Import buttons stay on the command bar (not the grid card) and
delegate to public panel methods (ExportCsv/ExportJson/Import), mirroring
how QueryEditorPanel exposes FocusEditor/OpenSearch. F6 focus hand-off uses
the panel's FocusGrid()/IsGridFocused; Escape-closes-inspector stays in
MainWindow's OnKeyDown (CellInspector lives on MainViewModel). Explain and
the staged-changes Review/Discard status-bar handlers stay in MainWindow,
rewritten to read _viewModel.ActiveTab instead of the removed field.

No behavior change, with one deliberate note: the cell inspector overlay
now lives inside the panel (at the editor/results split's results row)
rather than at the window root, so its scrim and centered card scope to the
results pane instead of the whole window.

MainWindow.axaml.cs: 2281 -> 1223 lines; MainWindow.axaml: 833 -> 589.
